backend/pyproject.toml
[project]
name = "raf-coffee-backend"
version = "0.1.0"
description = "Backend for coffee shop"
authors = [
    {name = "Rofar02",email = "lexa95k@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.136.0,<0.137.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "redis (>=7.4.0,<8.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "mypy (>=1.20.1,<2.0.0)"
]
